libusermode: make injections optional
