+++ configure.ac
@@ -545,6 +545,13 @@ AC_ARG_ENABLE([debug],
   [debug="no"])
 AM_CONDITIONAL([DEBUG], [test x$debug = xyes])
 
+AC_ARG_ENABLE([libusermode-injections],
+  [AS_HELP_STRING([--enable-libusermode-injections],
+    [Use injections instead of page faults in libusermode @<:@no@:>@])],
+  [libusermode-injections="$enableval"],
+  [libusermode-injections="no"])
+AM_CONDITIONAL([LIBUSERMODE_USE_INJECTION], [test x$libusermode-injections = xyes])
+
 AC_ARG_ENABLE([sanitize],
   [AS_HELP_STRING([--enable-sanitize],
     [Enable sanitizers to be compiled @<:@no@:>@])],
