+++ src/libusermode/meson.build
@@ -3,6 +3,8 @@ libusermode_h = include_directories(
 )
 libusermode = static_library('usermode',
     'userhook.cpp',
+    'userhook_pf.cpp',
+    'userhook_inject.cpp',
     'running.cpp',
     'printers/printers.cpp',
     'printers/utils.cpp',
